backend/requirements.txt
fastapi==0.115.6
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
uvicorn==0.34.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
bcrypt==4.0.1
python-dotenv==1.0.1
